Common logic extraction
