+++ src/commonMain/kotlin/operations/logic/unwrap/SingleNestedValueUnwrapStrategy.kt
@@ -1,0 +1,15 @@
+package operations.logic.unwrap
+
+// TODO requires unit tests
+internal interface SingleNestedValueUnwrapStrategy {
+    fun unwrapSingleNestedValueOrDefault(value: Any?) = value.unwrapSingleNestedValue().let {
+        if (it != value) {
+            SingleNestedValue(it)
+        } else value
+    }
+
+    private fun Any?.unwrapSingleNestedValue(): Any? = when {
+        this is List<*> && this.size == 1 -> this.firstOrNull().unwrapSingleNestedValue()
+        else -> this
+    }
+}
